support vector classifier+ postprocessing+ EDA
added a svc model, and fixed the ids problem with the postprocessor, allowing to write predictions to an external file. Also redid the embeddings with the stella model in google colab, to accomodate new data. Also updated the data analysis note book with some new nice plots for the EDA